Move pd_read/pd_write to utils to stop a circular dependency
